requirements.txt
appdirs==1.4.3
atomicwrites==1.3.0
attrs==18.2.0
black==18.9b0
Click==7.0
Django==2.1.4
djangorestframework==3.9.0
entrypoints==0.3
Faker==0.9.1
flake8==3.7.7
mccabe==0.6.1
mixer==6.1.3
more-itertools==6.0.0
pluggy==0.9.0
py==1.8.0
pycodestyle==2.5.0
pyflakes==2.1.1
pytest==4.3.0
python-dateutil==2.8.0
pytz==2018.7
six==1.12.0
text-unidecode==1.2
toml==0.10.0
